+++ packages/browser-wallet/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 -   Added export activity button to Transaction Log as part of feature parity with Old UI
+-   Added Allowlist config at account page as part of feature parity with Old UI
 
 ## 2.9.0